pyproject.toml
[project]
name = "bilibili-downloader"
version = "1.9"
description = "B站视频解析下载工具"
authors = [
    { name = "寒烟似雪", email = "2273962061@qq.com" }
]
license = { text = "MIT" }
requires-python = ">=3.6"
dependencies = [
    "PyQt5>=5.15.10",
    "requests>=2.31.0",
    "aiohttp>=3.9.3",
    "orjson>=3.10.0",
    "cachetools>=5.3.3",
    "python-dateutil>=2.9.0",
    "pycryptodome>=3.20.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ["py36"]

[tool.isort]
profile = "black"
